pyproject.toml
[project]
name = "foundation-llm-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "datasets>=3.6.0",
    "matplotlib>=3.10.5",
    "numpy>=2.2.4",
    "pyyaml>=6.0.2",
    "tiktoken>=0.9.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "wandb>=0.19.11",
]
